Bundle rate support, upload of images, better administrator notifications, minor bug fixing.
Support of bundle rates, and rates working on a group of calls in a certain time-frame.
New rate methods: cap limit, minimum monthly cost.
Improved management of dates in SQL queries.
Minor bug-fixings on filters, reset of calls and so on.
Better directory structure.
Administrator receives details about errors, and not only a notification of theirs presence.